requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
numpy>=1.26
scikit-learn>=1.4
scipy>=1.12
opencv-python>=4.9
Pillow>=10.2
tqdm>=4.66
matplotlib>=3.8
pyyaml>=6.0.1
pandas>=2.2
Flask>=2.3.0
PyMuPDF>=1.23.0
reportlab>=4.0.0
Werkzeug>=2.3.0
albumentations>=1.3.0
wandb>=0.16.0
